requirements.txt
Flask==3.1.0
gunicorn==23.0.0
werkzeug==3.1.3
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
reportlab>=4.0.0
